requirements.txt
Django==4.2
psycopg2-binary==2.9.6
djangorestframework==3.14
djangorestframework-simplejwt[crypto]==5.3.0